Log While Waiting For Record to Add to Batch
Output a log message when we're waiting for a record to add to a batch.

When the expression needs to process records to determine whether
updates are needed, it may rarely or never send records back to
Salesforce so we include some output to make it obvious that the
application is still working.