Abort On Panic In Converter
Allow the converter to cancel the query when a panic occurs.
Previously, the converter would drain all of the records coming from the
query before returning an error, which could take a long time.

Upgrade force library for fix to bug that could cause a send on a closed
channel when the query's context is cancelled.

Upgrade expr to the latest version.