Use Tooling API To Execute Anonymous Apex
When getting context variables use the Tooling API to execute anonymous
apex.

Because /tooling/executeAnonymous doesn't return the debug log
like the SOAP version, we use another approach to get the context
variables.  We throw an exception at the end of execution, setting the
exception message to the serialized context.